@@ -10,36 +10,36 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~12.1.0",
-    "@angular/common": "~12.1.0",
-    "@angular/compiler": "~12.1.0",
-    "@angular/core": "~12.1.0",
-    "@angular/forms": "~12.1.0",
-    "@angular/localize": "~12.1.0",
-    "@angular/platform-browser": "~12.1.0",
-    "@angular/platform-browser-dynamic": "~12.1.0",
-    "@angular/router": "~12.1.0",
-    "@ng-bootstrap/ng-bootstrap": "^11.0.0-beta.2",
-    "bootstrap": "^5.0.0-beta3",
-    "dotenv": "^16.0.0",
-    "gsap": "^3.7.0",
+    "@angular/animations": "^15.1.2",
+    "@angular/common": "^15.1.2",
+    "@angular/compiler": "^15.1.2",
+    "@angular/core": "^15.1.2",
+    "@angular/forms": "^15.1.2",
+    "@angular/localize": "^15.1.2",
+    "@angular/platform-browser": "^15.1.2",
+    "@angular/platform-browser-dynamic": "^15.1.2",
+    "@angular/router": "^15.1.2",
+    "@ng-bootstrap/ng-bootstrap": "^11.0.1",
+    "bootstrap": "^5.2.3",
+    "dotenv": "^16.0.3",
+    "gsap": "^3.11.4",
     "rxjs": "~6.6.0",
-    "tslib": "^2.2.0",
-    "zone.js": "~0.11.4"
+    "tslib": "^2.4.1",
+    "zone.js": "^0.11.8"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~12.1.0",
-    "@angular/cli": "~12.1.0",
-    "@angular/compiler-cli": "~12.1.0",
+    "@angular-devkit/build-angular": "^15.1.3",
+    "@angular/cli": "^15.1.3",
+    "@angular/compiler-cli": "^15.1.2",
     "@types/gsap": "^3.0.0",
     "@types/jasmine": "~3.6.0",
-    "@types/node": "^12.11.1",
+    "@types/node": "^12.20.55",
     "jasmine-core": "~3.7.0",
-    "karma": "^6.3.20",
-    "karma-chrome-launcher": "~3.1.0",
+    "karma": "^6.4.1",
+    "karma-chrome-launcher": "^3.1.1",
     "karma-coverage": "~2.0.3",
-    "karma-jasmine": "~4.0.0",
-    "karma-jasmine-html-reporter": "^1.5.0",
-    "typescript": "~4.3.2"
+    "karma-jasmine": "^4.0.2",
+    "karma-jasmine-html-reporter": "^1.7.0",
+    "typescript": "~4.9.4"
   }
 }
